refactor(frontend): 将 WebSocketProvider 重命名为 NetworkServiceProvider 并清理向后兼容别名
WebSocket 迁移完成后，文件名和导出名仍保留旧的 WebSocket 命名。
统一重命名为 NetworkServiceProvider，移除 WebSocketProvider 和
useWebSocketActions 的向后兼容别名，更新所有引用（11 处）。
